Add files via upload
First create a Web Api on Azure. Then create “.cs” files on C#. After that take an mongoDB account and copy the link from there to use it in “DB.cs" file to reach database. Create a file on MondDB to use on testing. By using “.Net" connect the web page and see the verifying. For testing use Postman and apply commands like get, post,delete and the others.